Visualize brain vs breast tissue prediction
